Move TestContext to dedicated module
